requirements.txt
biopython>=1.78
pandas>=1.3
matplotlib>=3.4
scipy>=1.7
scikit-posthocs>=0.6
